Move user activity utils into lib
